refactor of tib test
